backend/requirements.txt
django
djangorestframework
scrapy
pymongo
